Center the information flex items
When text wraps, the text no longer is centered under flex properties of
justify-content: center;
text-align: center; is also required.

The images flex container is centered. However, as soon as one of the
images wraps, the images appear to be off center slightly. At that point
it is necessary to also justify-content: center; the images flex
container so the images remain centered when wrapping.